fix: error handling
